requirements.txt
openai>=1.40.0
anthropic>=0.39.0
ollama>=0.3.0
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.120.0
duckduckgo-search>=6.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
rich>=13.7.0
gradio>=4.0.0
